requirements.txt
# Core dependencies
openai==1.35.3
streamlit==1.35.0
pymupdf==1.24.5
chromadb==0.5.0
python-dotenv==1.0.1
tiktoken==0.7.0
protobuf==3.20.3  # Fix for ChromaDB compatibility

# Optional but recommended
langchain==0.2.5
langchain-openai==0.1.9
langchain-community==0.2.5

# Utilities
numpy==1.26.4
pydantic==2.7.4
tenacity==8.4.1
